refactor(core): co-locate View and Region ownership
